Use Form Objects to handle validations
**Why**: Since we allow the user to create their account in multiple
steps, some validations don't apply all the time. Instead of adding
conditional logic and flags to the User model, it's cleaner to separate
responsibilities by using Form Objects for each form that only include
the validations that apply to that form.

For example, we know that when a user enters their email in the first
step, the only thing we need to validate is their email.

To make it easier to test these new forms, I updated some of the
Active Job jobs to take an email or mobile as a parameter, instead of
a User object. Strings are preferred to objects as parameters so that
you don't have to worry about deserialization errors.