Add script + config to quickly generate new SAML PKI
__Why__
We may need the ability to quickly generate and rotate keys at some point in the
future. This script + config provides a rudimentary solution for that need.

__How__
Create a script to invoke the openssl binary to generate a new encrypted key and
IdP cert using the config file at conf/openssl.conf (thanks @sbc100).