Switch SmsSenderOtpJob to take PORO rather than User model
**Why**: Removes the need for sidekiq to do database access.
This matches the behaviour of the other jobs.  Also allows
us the remove some redunant specs.  It also moves the logic
for deciding which number to send to out of async job which
is required another change I'm working to store unconfirmed
mobile number in the session rather than in the database.